Extract TableLayoutSettings classes
