+++ src/System.Windows.Forms/src/System/Windows/Forms/TableLayoutSettings.cs
@@ -12,10 +12,12 @@ using System.Windows.Forms.Layout;
 
 namespace System.Windows.Forms
 {
-    /// <summary>this is a wrapper class to expose interesting properties of TableLayout</summary>
+    /// <summary>
+    ///  This is a wrapper class to expose interesting properties of TableLayout
+    /// </summary>
     [TypeConverter(typeof(TableLayoutSettingsTypeConverter))]
     [Serializable]  // This class participates in resx serialization.
-    public sealed class TableLayoutSettings : LayoutSettings, ISerializable
+    public sealed partial class TableLayoutSettings : LayoutSettings, ISerializable
     {
         private static readonly int[] borderStyleToOffset = {
             /*None = */ 0,
@@ -565,302 +567,6 @@ namespace System.Windows.Forms
                 }
                 return controlsInfo;
             }
-
-        }
-
-        internal struct ControlInformation
-        {
-            internal object Name;
-            internal int Row;
-            internal int Column;
-            internal int RowSpan;
-            internal int ColumnSpan;
-
-            internal ControlInformation(object name, int row, int column, int rowSpan, int columnSpan)
-            {
-                Name = name;
-                Row = row;
-                Column = column;
-                RowSpan = rowSpan;
-                ColumnSpan = columnSpan;
-            }
-        }
-
-        /// <summary> TableLayoutSettingsStub
-        ///     contains information about
-        /// </summary>
-        private class TableLayoutSettingsStub
-        {
-            private static ControlInformation DefaultControlInfo = new ControlInformation(null, -1, -1, 1, 1);
-            private TableLayoutColumnStyleCollection columnStyles;
-            private TableLayoutRowStyleCollection rowStyles;
-            private Dictionary<object, ControlInformation> controlsInfo;
-
-            public TableLayoutSettingsStub()
-            {
-            }
-
-            /// <summary> ApplySettings - applies settings from the stub into a full-fledged
-            ///  TableLayoutSettings.
-            ///
-            ///  NOTE: this is a one-time only operation - there is data loss to the stub
-            ///  as a result of calling this function. we hand as much over to the other guy
-            ///  so we dont have to reallocate anything
-            /// </summary>
-            internal void ApplySettings(TableLayoutSettings settings)
-            {
-                //
-                // apply row,column,rowspan,colspan
-                //
-                TableLayout.ContainerInfo containerInfo = TableLayout.GetContainerInfo(settings.Owner);
-                if (containerInfo.Container is Control appliedControl && controlsInfo != null)
-                {
-
-                    // we store the control names, look up the controls
-                    // in the appliedControl's control collection and apply the row,column settings.
-                    foreach (object controlName in controlsInfo.Keys)
-                    {
-                        ControlInformation controlInfo = controlsInfo[controlName];
-
-                        // Look for the control in our table, we have to go through
-                        // PropertyDescriptor rather than just going using appliedControl.Controls[controlName]
-                        // because the Name property is shadowed at design time
-                        foreach (Control tableControl in appliedControl.Controls)
-                        {
-                            if (tableControl != null)
-                            {
-                                string name = null;
-                                PropertyDescriptor prop = TypeDescriptor.GetProperties(tableControl)["Name"];
-                                if (prop != null && prop.PropertyType == typeof(string))
-                                {
-                                    name = prop.GetValue(tableControl) as string;
-                                }
-
-                                if (WindowsFormsUtils.SafeCompareStrings(name, controlName as string, /* ignoreCase = */ false))
-                                {
-                                    settings.SetRow(tableControl, controlInfo.Row);
-                                    settings.SetColumn(tableControl, controlInfo.Column);
-                                    settings.SetRowSpan(tableControl, controlInfo.RowSpan);
-                                    settings.SetColumnSpan(tableControl, controlInfo.ColumnSpan);
-                                    break;
-                                }
-                            }
-                        }
-                    }
-                }
-
-                //
-                // assign over the row and column styles
-                //
-                containerInfo.RowStyles = rowStyles;
-                containerInfo.ColumnStyles = columnStyles;
-
-                // since we've given over the styles to the other guy, null out.
-                columnStyles = null;
-                rowStyles = null;
-            }
-
-            public TableLayoutColumnStyleCollection ColumnStyles
-            {
-                get
-                {
-                    if (columnStyles == null)
-                    {
-                        columnStyles = new TableLayoutColumnStyleCollection();
-                    }
-                    return columnStyles;
-                }
-            }
-
-            public TableLayoutRowStyleCollection RowStyles
-            {
-                get
-                {
-                    if (rowStyles == null)
-                    {
-                        rowStyles = new TableLayoutRowStyleCollection();
-                    }
-                    return rowStyles;
-                }
-            }
-
-            internal List<ControlInformation> GetControlsInformation()
-            {
-
-                if (controlsInfo == null)
-                {
-                    return new List<ControlInformation>();
-                }
-
-                List<ControlInformation> listOfControlInfo = new List<ControlInformation>(controlsInfo.Count);
-                foreach (object name in controlsInfo.Keys)
-                {
-                    ControlInformation ci = controlsInfo[name];
-                    ci.Name = name;
-                    listOfControlInfo.Add(ci);
-                }
-                return listOfControlInfo;
-            }
-
-            private ControlInformation GetControlInformation(object controlName)
-            {
-
-                if (controlsInfo == null)
-                {
-                    return DefaultControlInfo;
-                }
-                if (!controlsInfo.ContainsKey(controlName))
-                {
-                    return DefaultControlInfo;
-                }
-                return controlsInfo[controlName];
-
-            }
-
-            public int GetColumn(object controlName)
-            {
-                return GetControlInformation(controlName).Column;
-            }
-            public int GetColumnSpan(object controlName)
-            {
-                return GetControlInformation(controlName).ColumnSpan;
-            }
-            public int GetRow(object controlName)
-            {
-                return GetControlInformation(controlName).Row;
-            }
-            public int GetRowSpan(object controlName)
-            {
-                return GetControlInformation(controlName).RowSpan;
-            }
-
-            private void SetControlInformation(object controlName, ControlInformation info)
-            {
-                if (controlsInfo == null)
-                {
-                    controlsInfo = new Dictionary<object, ControlInformation>();
-                }
-                controlsInfo[controlName] = info;
-            }
-
-            public void SetColumn(object controlName, int column)
-            {
-                if (GetColumn(controlName) != column)
-                {
-                    ControlInformation info = GetControlInformation(controlName);
-                    info.Column = column;
-                    SetControlInformation(controlName, info);
-                }
-
-            }
-            public void SetColumnSpan(object controlName, int value)
-            {
-                if (GetColumnSpan(controlName) != value)
-                {
-                    ControlInformation info = GetControlInformation(controlName);
-                    info.ColumnSpan = value;
-                    SetControlInformation(controlName, info);
-                }
-            }
-            public void SetRow(object controlName, int row)
-            {
-                if (GetRow(controlName) != row)
-                {
-                    ControlInformation info = GetControlInformation(controlName);
-                    info.Row = row;
-                    SetControlInformation(controlName, info);
-                }
-            }
-            public void SetRowSpan(object controlName, int value)
-            {
-                if (GetRowSpan(controlName) != value)
-                {
-                    ControlInformation info = GetControlInformation(controlName);
-                    info.RowSpan = value;
-                    SetControlInformation(controlName, info);
-                }
-            }
-
-        } // end of System.Windows.Forms.TableLayoutSettings
-
-        internal class StyleConverter : TypeConverter
-        {
-            public override bool CanConvertTo(ITypeDescriptorContext context, Type destinationType)
-            {
-                if (destinationType == typeof(InstanceDescriptor))
-                {
-                    return true;
-                }
-                return base.CanConvertTo(context, destinationType);
-            }
-
-            public override object ConvertTo(ITypeDescriptorContext context, CultureInfo culture, object value, Type destinationType)
-            {
-                if (destinationType == typeof(InstanceDescriptor) && value is TableLayoutStyle)
-                {
-                    TableLayoutStyle style = (TableLayoutStyle)value;
-
-                    switch (style.SizeType)
-                    {
-                        case SizeType.AutoSize:
-                            return new InstanceDescriptor(
-                                style.GetType().GetConstructor(new Type[] { }),
-                                new object[] { });
-                        case SizeType.Absolute:
-                        case SizeType.Percent:
-                            return new InstanceDescriptor(
-                                style.GetType().GetConstructor(new Type[] { typeof(SizeType), typeof(int) }),
-                                new object[] { style.SizeType, style.Size });
-                        default:
-                            break;
-                    }
-                }
-                return base.ConvertTo(context, culture, value, destinationType);
-            }
-        }
-    }
-
-    public class ColumnStyle : TableLayoutStyle
-    {
-        public ColumnStyle() { }
-
-        public ColumnStyle(SizeType sizeType)
-        {
-            SizeType = sizeType;
-        }
-
-        public ColumnStyle(SizeType sizeType, float width)
-        {
-            SizeType = sizeType;
-            Width = width;
-        }
-
-        public float Width
-        {
-            get { return base.Size; }
-            set { base.Size = value; }
-        }
-    }
-
-    public class RowStyle : TableLayoutStyle
-    {
-        public RowStyle() { }
-
-        public RowStyle(SizeType sizeType)
-        {
-            SizeType = sizeType;
-        }
-
-        public RowStyle(SizeType sizeType, float height)
-        {
-            SizeType = sizeType;
-            Height = height;
-        }
-
-        public float Height
-        {
-            get { return base.Size; }
-            set { base.Size = value; }
         }
     }
 }
